# Use single thread per request to reduce mutex contention issues under load