--- a/backend/mse/card_str_match.c
+++ b/backend/mse/card_str_match.c
@@ -156,17 +156,9 @@ static int __mse_match_card_node(mse_avl_tree_node_t * restrict node,
         return 1;
     }
 
-    if (!__mse_match_card_do_match(node, match_data, cmp_data)) {
-        return 0;
-    }
-
-    if (!__mse_match_card_node(node->l, match_data, cmp_data)) {
-        return 0;
-    }
-
-    if (!__mse_match_card_node(node->r, match_data, cmp_data)) {
-        return 0;
-    }
+    ASSERT(__mse_match_card_do_match(node, match_data, cmp_data));
+    ASSERT(__mse_match_card_node(node->l, match_data, cmp_data));
+    ASSERT(__mse_match_card_node(node->r, match_data, cmp_data));
 
     return 1;
 }