--- a/backend/mse/consumer.h
+++ b/backend/mse/consumer.h
@@ -1,7 +1,6 @@
 #pragma once
 #include <stddef.h>
 #include "./avl_tree.h"
-#include "./thread_pool.h"
 #include "./mtg_json.h"
 #include "./search.h"
 
@@ -33,5 +32,4 @@ void mse_free_set_consumer(mse_set_consumer_t *gen);
 int mse_consume_set(mse_set_consumer_t *gen,
                     mse_search_intermediate_t *res,
                     mse_all_printings_cards_t *cards,
-                    mse_search_intermediate_t *child,
-                    mse_thread_pool_t *pool);
+                    mse_search_intermediate_t *child);