--- a/backend/mse/generator_name.h
+++ b/backend/mse/generator_name.h
@@ -1,10 +1,8 @@
 #pragma once
 #include "./generators.h"
 #include "./avl_tree.h"
-#include "./thread_pool.h"
 #include "./mtg_json.h"
 
 int mse_generate_set_name(mse_set_generator_t *gen,
                           mse_search_intermediate_t *res,
-                          mse_all_printings_cards_t *cards,
-                          mse_thread_pool_t *pool);
+                          mse_all_printings_cards_t *cards);