--- a/backend/mse/generator_oracle.c
+++ b/backend/mse/generator_oracle.c
@@ -5,14 +5,13 @@
 
 static int __mse_generate_set_oracle_re(mse_set_generator_t *gen,
                                         mse_search_intermediate_t *res,
-                                        mse_all_printings_cards_t *cards,
-                                        mse_thread_pool_t *pool)
+                                        mse_all_printings_cards_t *cards)
 {
     char *re = mse_escape_regex(gen->argument);
     ASSERT(re != NULL);
 
     mse_avl_tree_node_t *node = NULL;
-    int status = mse_matching_card_oracle(&node, cards->card_tree, re, 1, gen->negate, pool);
+    int status = mse_matching_card_oracle(&node, cards->card_tree, re, 1, gen->negate);
     *res = mse_init_search_intermediate_tree(node, 0);
     free(re);
 
@@ -22,23 +21,21 @@ static int __mse_generate_set_oracle_re(mse_set_generator_t *gen,
 
 static int __mse_generate_set_oracle_text_inc(mse_set_generator_t *gen,
         mse_search_intermediate_t *res,
-        mse_all_printings_cards_t *cards,
-        mse_thread_pool_t *pool)
+        mse_all_printings_cards_t *cards)
 {
     mse_avl_tree_node_t *node = NULL;
-    ASSERT(mse_matching_card_oracle(&node, cards->card_tree, gen->argument, 0, gen->negate, pool));
+    ASSERT(mse_matching_card_oracle(&node, cards->card_tree, gen->argument, 0, gen->negate));
     *res = mse_init_search_intermediate_tree(node, 0);
     return 1;
 }
 
 int mse_generate_set_oracle(mse_set_generator_t *gen,
                             mse_search_intermediate_t *res,
-                            mse_all_printings_cards_t *cards,
-                            mse_thread_pool_t *pool)
+                            mse_all_printings_cards_t *cards)
 {
     if (mse_is_regex_str(gen->argument)) {
-        return __mse_generate_set_oracle_re(gen, res, cards, pool);
+        return __mse_generate_set_oracle_re(gen, res, cards);
     } else {
-        return __mse_generate_set_oracle_text_inc(gen, res, cards, pool);
+        return __mse_generate_set_oracle_text_inc(gen, res, cards);
     }
 }