--- a/backend/mse/generators.h
+++ b/backend/mse/generators.h
@@ -1,7 +1,6 @@
 #pragma once
 #include <stddef.h>
 #include "./avl_tree.h"
-#include "./thread_pool.h"
 #include "./mtg_json.h"
 #include "./search.h"
 
@@ -70,8 +69,7 @@ int __mse_validate_generator_op_combo(mse_set_generator_type_t gen_type,
 
 int mse_generate_set(mse_set_generator_t *gen,
                      mse_search_intermediate_t *res,
-                     mse_all_printings_cards_t *cards,
-                     mse_thread_pool_t *pool);
+                     mse_all_printings_cards_t *cards);
 
 /// String to generator type
 int mse_gen_type(char *str, mse_set_generator_type_t *ret);