--- a/backend/tests/test_generator_colours.c
+++ b/backend/tests/test_generator_colours.c
@@ -32,12 +32,12 @@ static int test_generator_colours_##op_l() \
     ASSERT(mse_init_set_generator(&ret, gen_type, MSE_SET_GENERATOR_OP_##op_u, DEFAULT_ARGUMENT_STR, len)); \
  \
     mse_search_intermediate_t inter; \
-    ASSERT(mse_generate_set(&ret, &inter, &gen_cards, &gen_thread_pool)); \
+    ASSERT(mse_generate_set(&ret, &inter, &gen_cards)); \
     ASSERT(mse_tree_size(inter.node) > 0); \
     ASSERT(test_tree_c_##op_l(inter.node, ret.negate)); \
  \
     ret.negate = 1; \
-    ASSERT(mse_generate_set(&ret, &inter, &gen_cards, &gen_thread_pool)); \
+    ASSERT(mse_generate_set(&ret, &inter, &gen_cards)); \
     ASSERT(mse_tree_size(inter.node) > 0); \
     ASSERT(test_tree_c_##op_l(inter.node, ret.negate)); \
  \
@@ -77,12 +77,12 @@ static int test_generator_colour_identity_##op_l() \
     ASSERT(mse_init_set_generator(&ret, gen_type, MSE_SET_GENERATOR_OP_##op_u, DEFAULT_ARGUMENT_STR, len)); \
  \
     mse_search_intermediate_t inter; \
-    ASSERT(mse_generate_set(&ret, &inter, &gen_cards, &gen_thread_pool)); \
+    ASSERT(mse_generate_set(&ret, &inter, &gen_cards)); \
     ASSERT(mse_tree_size(inter.node) > 0); \
     ASSERT(test_tree_ci_##op_l(inter.node, ret.negate)); \
  \
     ret.negate = 1; \
-    ASSERT(mse_generate_set(&ret, &inter, &gen_cards, &gen_thread_pool)); \
+    ASSERT(mse_generate_set(&ret, &inter, &gen_cards)); \
     ASSERT(mse_tree_size(inter.node) > 0); \
     ASSERT(test_tree_ci_##op_l(inter.node, ret.negate)); \
     mse_free_search_intermediate(&inter); \