This will make some cases slower, but extreme load (the most exciting case) faster.