@djpiper28 2025-08-17
# Before Changes

```
hey -n 1000 -c 50 -m GET -d "type:goblin and name:krenko and type:legendary and o:/.*craete.*tokens.*/" http://localhost:4365/api

Summary:
  Total:        58.2840 secs
  Slowest:      12.8082 secs
  Fastest:      0.0608 secs
  Average:      2.7242 secs
  Requests/sec: 17.1574
  
  Total data:   59000 bytes
  Size/request: 59 bytes

Response time histogram:
  0.061 [1]     |
  1.336 [251]   |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  2.610 [318]   |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  3.885 [207]   |■■■■■■■■■■■■■■■■■■■■■■■■■■
  5.160 [121]   |■■■■■■■■■■■■■■■
  6.435 [42]    |■■■■■
  7.709 [31]    |■■■■
  8.984 [17]    |■■
  10.259 [7]    |■
  11.534 [4]    |■
  12.808 [1]    |


Latency distribution:
  10% in 0.6975 secs
  25% in 1.3296 secs
  50% in 2.2050 secs
  75% in 3.6855 secs
  90% in 5.1863 secs
  95% in 6.7474 secs
  99% in 9.0855 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0003 secs, 0.0608 secs, 12.8082 secs
  DNS-lookup:   0.0001 secs, 0.0000 secs, 0.0063 secs
  req write:    0.0001 secs, 0.0000 secs, 0.0049 secs
  resp wait:    2.7236 secs, 0.0607 secs, 12.8081 secs
  resp read:    0.0001 secs, 0.0000 secs, 0.0020 secs

Status code distribution:
  [200] 1000 responses

```


# After Changes

```
hey -n 1000 -c 50 -m GET -d "type:goblin and name:krenko and type:legendary and o:/.*craete.*tokens.*/" http://localhost:4365/api

Summary:
  Total:        0.2023 secs
  Slowest:      0.0305 secs
  Fastest:      0.0060 secs
  Average:      0.0093 secs
  Requests/sec: 4943.8663
  
  Total data:   59000 bytes
  Size/request: 59 bytes

Response time histogram:
  0.006 [1]     |
  0.008 [438]   |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.011 [463]   |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.013 [32]    |■■■
  0.016 [10]    |■
  0.018 [13]    |■
  0.021 [29]    |■■■
  0.023 [4]     |
  0.026 [3]     |
  0.028 [5]     |
  0.030 [2]     |


Latency distribution:
  10% in 0.0073 secs
  25% in 0.0079 secs
  50% in 0.0086 secs
  75% in 0.0093 secs
  90% in 0.0108 secs
  95% in 0.0172 secs
  99% in 0.0231 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0003 secs, 0.0060 secs, 0.0305 secs
  DNS-lookup:   0.0001 secs, 0.0000 secs, 0.0070 secs
  req write:    0.0001 secs, 0.0000 secs, 0.0031 secs
  resp wait:    0.0087 secs, 0.0060 secs, 0.0234 secs
  resp read:    0.0001 secs, 0.0000 secs, 0.0035 secs

Status code distribution:
  [200] 1000 responses
```