@djpiper28 2025-08-17
on a larger scale:

```
mtg-search-engine/bench on  main [$!] via  v3.12.8 (.venv) via  impure (nix-shell) 
❯ hey -n 1000000 -c 500 -m GET -d "type:goblin and name:krenko and type:legendary and o:/.*craete.*tokens.*/" http://localhost:4365/api

Summary:
  Total:        139.9064 secs
  Slowest:      0.3206 secs
  Fastest:      0.0047 secs
  Average:      0.0698 secs
  Requests/sec: 7147.6335
  
  Total data:   59000000 bytes
  Size/request: 59 bytes

Response time histogram:
  0.005 [1]     |
  0.036 [6600]  |■
  0.068 [439617]        |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.099 [526918]        |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.131 [26529] |■■
  0.163 [113]   |
  0.194 [47]    |
  0.226 [89]    |
  0.257 [12]    |
  0.289 [11]    |
  0.321 [63]    |


Latency distribution:
  10% in 0.0526 secs
  25% in 0.0617 secs
  50% in 0.0691 secs
  75% in 0.0773 secs
  90% in 0.0875 secs
  95% in 0.0942 secs
  99% in 0.1074 secs

Details (average, fastest, slowest):
  DNS+dialup:   0.0000 secs, 0.0047 secs, 0.3206 secs
  DNS-lookup:   0.0000 secs, 0.0000 secs, 0.0317 secs
  req write:    0.0000 secs, 0.0000 secs, 0.0320 secs
  resp wait:    0.0696 secs, 0.0046 secs, 0.2850 secs
  resp read:    0.0001 secs, 0.0000 secs, 0.0484 secs

Status code distribution:
  [200] 1000000 responses
```